refactored initialisers to use gcd

+++ src/Catty/Scene.m
@@ -29,7 +29,6 @@
 
 
 @interface Scene()
-
 @property (nonatomic) NSTimeInterval lastUpdateTimeInterval;
 
 @end
@@ -37,7 +36,7 @@
 
 @implementation Scene
 
-- (id) initWithSize:(CGSize)size andProgram:(Program *)program
+- (id)initWithSize:(CGSize)size andProgram:(Program *)program
 {
     if (self = [super initWithSize:size]) {
         self.program = program;
@@ -85,31 +84,34 @@
     return scenePoint;
 }
 
--(float)convertYCoordinateToScene:(float)y {
+-(CGFloat)convertYCoordinateToScene:(CGFloat)y
+{
     return (self.size.height/2.0f + y);
 }
 
--(float)convertXCoordinateToScene:(float)x {
+-(CGFloat)convertXCoordinateToScene:(CGFloat)x
+{
     return (self.scene.size.width/2.0f + x);
 }
 
 -(CGPoint)convertSceneCoordinateToPoint:(CGPoint)point
 {
-    float x = point.x - self.scene.size.width/2.0f;
-    float y = point.y - self.scene.size.height/2.0f;
+    CGFloat x = point.x - self.scene.size.width/2.0f;
+    CGFloat y = point.y - self.scene.size.height/2.0f;
     return CGPointMake(x, y);
 }
 
 -(CGFloat) convertDegreesToScene:(CGFloat)degrees
 {
-    return 360.0 - degrees;
+    return 360.0f - degrees;
 }
 
 -(CGFloat) convertSceneToDegrees:(CGFloat)degrees
 {
-    return 360.0 + degrees;
+    return 360.0f + degrees;
 }
--(BOOL)touchedwith:(NSSet*)touches withX:(CGFloat) x andY:(CGFloat) y
+
+-(BOOL)touchedwith:(NSSet*)touches withX:(CGFloat)x andY:(CGFloat)y
 {
     NSDebug(@"StartTouchofScene");
     UITouch *touch = [touches anyObject];
